+++ back-end/hospital-api/src/main/java/net/pladema/provincialreports/programreports/repository/EpidemiologyOneConsultationDetail.java
@@ -1,4 +1,4 @@
-package net.pladema.programreports.repository;
+package net.pladema.provincialreports.programreports.repository;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
